Remove state router
